Added ability to build on test and working on running commands in container
Use strider-docker-runner to run commands.

Changed to dropdown and removed IIFEs

Forgot about how picky Angular is with the selected attribute

Changed verbiage
