+++ config/config.html
@@ -3,8 +3,18 @@
 <hr>
 <div class="row-fluid">
   <div class="span12">
-    <label for="tagName">Tag name</label>
-    <input type="text" name="tagName" ng-model="config.tag">
+    <label for="buildPhase">On which phase would you like to build your image?</label>
+    <select name="buildPhase" ng-model="config.buildPhase" required>
+      <option value="deploy" ng-selected="!config.buildPhase || config.buildPhase == 'deploy'">Deploy</option>
+      <option value="test" ng-selected="config.buildPhase == 'test">Test</option>
+    </select>
   </div>
 </div>
+<div class="row-fluid">
+  <div class="span12">
+    <label for="tagName">Tag name</label>
+    <input type="text" name="tagName" ng-model="config.tag" required>
+  </div>
+</div>
+
 <button ng-click="save()" class="btn btn-primary">Save</button>
